Income Tax Calculator
Write a program called IncomeTaxCalculator that reads the taxable income (in int). The program shall calculate the income tax payable (in double) and print the result. The progressive income tax table is mandated as follows
Taxable Income = First 20,000 is 0% rate, Taxable Income = Next 20,000 is 10% rate, Taxable Income = Next 20,000 is 20% rate, Taxable Income = Remaining 20,000 is 30% rate. For example, suppose that the taxable income is 85,000. The income tax payable is 20,000*0%+20,000*1%+20,000*2%+20,2000*3%